Switch footstep planning to terrain-aware mode.
The footstep planner now interacts with the heightmap to design a separate trajectory for each swing phase in the footstep plan. This allows the planner to automatically avoid terrain obstacles like stairs. The planner still uses the apex position given by the user as input, but adds the additional constraint that the swing trajectory must not cause the foot to come within a threshold of the terrain. 

Currently, this behavior is only correct for flat-footed walking, but it should perform acceptably on the ramp as well. 


git-svn-id: https://svn.csail.mit.edu/drc@3435 c7283977-0100-402a-a91a-fa70b306dbfe