Walking planner is now more transactional. Details:
Walking planner listens on APPROVED_FOOTSTEP_PLAN and COMMITTED_FOOTSTEP_PLAN. Messages to either result in robot plans being generated and sent to the viewer for review. Messages to COMMITTED_FOOTSTEP_PLAN additionally result in a walking_plan being sent over the WALKING_PLAN channel, to which the controller is listening.


git-svn-id: https://svn.csail.mit.edu/drc@3994 c7283977-0100-402a-a91a-fa70b306dbfe