Update footstep and walking planners for more sophisticated constraints.
The footstep planner now also specifies min heights for the toe and heel contacts to ensure that they do not drag on the ground. The footstep plan also has a higher default apex to help keep the steps from falling behind. In addition, we can now support fixing the pos of any link on the robot, for example fixing the hand while we walk. 

git-svn-id: https://svn.csail.mit.edu/drc@4403 c7283977-0100-402a-a91a-fa70b306dbfe