data fusion update request message has been put together,
VRC Leg Odometry code is tied in and seems operational.

There is a problem with forward kinematics not finding some of the joints -- this must be investigated.

After some updates there seems to be a problem with the parameter server as well. The model client does not find an active parameter update server, which used to work fine. Again, this must also be investigated

git-svn-id: https://svn.csail.mit.edu/drc/trunk@6486 c7283977-0100-402a-a91a-fa70b306dbfe