Add a very simple "Drake Controller Options" renderer to the viewer.
We've removed the map mode selector from the Data Control renderer, but our Drake controllers still rely on the MAP_CONTROLLER_COMMAND messages to know which heightmaps to use. I've separated this functionality off into a simple renderer which, for now, just lets the user set that parameter for any controllers which may be listening. This renderer does no input handling or drawing, so I'm hoping that integrating it will be low-risk.

git-svn-id: https://svn.csail.mit.edu/drc/trunk@7012 c7283977-0100-402a-a91a-fa70b306dbfe