bug 1632 - fix lcm-gen rules in lcmtypes.cmake
This is a partial fix related to bug 1632.

Running make reconfigure caused the next make to fail because the
lcm-gen command line that is run during configuration was missing the
c-path and cpp-hpath args.

This fix should go upstream and be applied to all the lcmtypes.cmake
files that are duplicated throughout the DRC repo and externals.  For
now, apply it just to drc_lcmtypes which is the most active lcmtypes
directory.

The lcmtypes.cmake file still has some incorrect generation rules that
need to be fixed in the future:

- it always regenerates files when cmake runs, triggering a full
  recompile of the lcmtypes library and all downstream client
  libraries using the headers, which is made much worse due to the
  "convenience" aggregate header

- it doesn't correctly handle lcm type files that are added/deleted

git-svn-id: https://svn.csail.mit.edu/drc/trunk@7034 c7283977-0100-402a-a91a-fa70b306dbfe