Pass terrain data with footstep plan, to allow automatic stepping over obstacles in BDI mode
Each footstep goal in a footsteps plan now specifies the simplified convex hull of the terrain from the prior step to that step. This has a couple of advantages: first, it allows us to pass the complete information needed to compute a walking plan in the footstep plan, rather than requiring additional calls to getHeight(); and second, it allows the BDI step translator to set the lift_height of steps in order to automatically clear obstacles in the terrain map.

git-svn-id: https://svn.csail.mit.edu/drc/trunk@7213 c7283977-0100-402a-a91a-fa70b306dbfe