remove scene height request from DRCTerrainMap constructor
The DRCTerrainMap constructor waited for a scene height in order to
record the minimum z value.  The min z value was used to fill in
heights where terrain maps had no samples.  This was a workaround and
now I am removing it.  I am setting minval to NaN and a follow up commit
by Robin is going to introduce a kinematic terrain map to replace the
NaNs.