alternate approach which allows for automatic trimming of steps
I'm playing around with a new idea here. The general concept is that
rather than brute-force searching over possible numbers of footsteps or
trying to add a quadratic cost on goal distance for every step, I can
instead add a binary variable for each step which indicates whether that
step is fixed to the goal pose or not. If I do this correctly, I get a
set of nominal-length footsteps up to the goal, and then a pile of fixed
steps at the goal, which I can trim out after the fact.