Temp hack fix for contact issues in drakeWalkingWithSensor
Some interaction of the new sensor data / changed outputFrame
causes noncontact-to-contact transition to not be detected
at a contact_threshold of 0.001, but it works fine at 0.01.
Still a problem is that the new sensor data (or introduction
of obstacles to the environment) causes upper body to drift
(at the back joint, I think) into more and more highly angled
orientations until Atlas just falls over. (This happens over
time as it walks along its planned path. Shouldn't there be
a feedback controller keeping that from happening???)