Add setEnvironment methods to IKServer and ik.py
The MATLAB method takes a urdf string and adds it to the robot RBM (which
contains just the robot itself) to produce the robot_and_environment RBM, which
is the one used for planning.

The python method takes in an lxml.ElementTree object, converts it to a string
and calls the MATLAB method.