remove remote/combined planner duplication by deriving from new DRCPlanner
CombinedPlanner() now behaves exactly as it did before we reorganized
for the split LCM communities. So, if you are operating entirely on one
machine or one community, CombinedPlanner() will just work.

For the split environment, we now have BasePlanner() and
RemotePlanner(), which live on the operator-side and controller-side,
respectively.

CombinedPlanner now just descends from BasePlanner and RemotePlanner and
subscribes to both sets of channels.