change scripts to wait for processes to start/end.  fixes #1233
Using killall -w to wait for process to end.
Using bash loop to wait for process to start.

Hopefully this guarantees that the script will succeed every time,
or hang visibly if there is some issue that prevents a process from
starting/ending cleanly.

Note, killall -w automatically waits at least 1 second if there is a
process to be killed.  So if you run the restart script multiple times
then there will be a 1 second delay per command.