simplify param set names
previously, drcAtlasParams.getDefaults returned a struct with fields
standing_sim, standing_hardware, walking_sim, walking_hardware, etc.
Plans would be labeled with a gain_set name like 'standing' or
'walking'. DRCPlanEval would be in charge of knowing if we were in sim
or hardware and appending the '_sim' or '_hardware' suffix to the param
set before sending it to the controller. This was needlessly complex.

Now, drcAtlasParams.getDefaults takes a use_sim_params flag, and returns
a struct with fields 'standing', 'walking', etc. which map to the
correct Sim/Hardware variant based on that flag. As a result,
DRCPlanEval no longer has to know or care if we're in sim. The only
place we ever need to record that information is in the single call to
getDefaults, which happens inside runDRCInstQP. This also makes the DRC
code more similar to our drake examples, since the
drake/examples/Atlas/atlasParams.getDefaults already uses 'standing',
'walking' etc. as its field names